# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Codeplane API
  version: 1.0.0
  description: API for Codeplane - a jj-native development platform
servers:
  - url: https://api.codeplane.app
paths:
...
  /repos/{owner}/{repo}/hooks:
    get:
      summary: List webhooks for a repository
      operationId: list_repo_hooks
      tags:
        - Webhooks
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          description: Opaque cursor from a previous response's next_cursor field
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
            maximum: 100
      responses:
        '200':
          description: List of webhooks
          headers:
            Link:
              description: RFC 8288 pagination links with cursor-based URLs.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
                  next_cursor:
                    type: string
                    nullable: true
                    description: Opaque cursor for fetching the next page. Null when there are no more results.
    post:
      summary: Create a webhook for a repository
      operationId: create_repo_hook
      tags:
        - Webhooks
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHookOption'
      responses:
        '201':
          description: Webhook created successfully. A ping event is sent to verify the endpoint.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '422':
          description: Validation error (invalid URL, missing events, etc.)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  errors:
                    type: array
                    items:
                      type: object
  /repos/{owner}/{repo}/hooks/{id}:
    get:
      summary: Get a webhook by ID
      operationId: get_repo_hook
      tags:
        - Webhooks
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Webhook details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '404':
          description: Webhook not found
    patch:
      summary: Update a webhook
      operationId: update_repo_hook
      tags:
        - Webhooks
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditHookOption'
      responses:
        '200':
          description: Webhook updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '404':
          description: Webhook not found
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  errors:
                    type: array
                    items:
                      type: object
    delete:
      summary: Delete a webhook
      operationId: delete_repo_hook
      tags:
        - Webhooks
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Webhook deleted successfully
        '404':
          description: Webhook not found
...
components:
  schemas:
...
    Webhook:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique webhook identifier
        type:
          type: string
          description: Webhook type (e.g., "codeplane", "slack", "discord")
          example: codeplane
        url:
          type: string
          format: uri
          description: The delivery URL
        config:
          type: object
          properties:
            content_type:
              type: string
              enum: [json, form]
              description: Payload content type
            url:
              type: string
              format: uri
              description: The delivery URL
          description: Webhook configuration
        events:
          type: array
          items:
            type: string
          description: List of subscribed event names
          example: ["push", "landing_request"]
        active:
          type: boolean
          description: Whether the webhook is active
        last_status:
          type: string
          enum: [none, succeed, fail]
          description: Status of the last delivery attempt
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateHookOption:
      type: object
      required:
        - type
        - config
        - events
      properties:
        type:
          type: string
          description: Webhook type
          example: codeplane
        config:
          type: object
          required:
            - url
            - content_type
          properties:
            url:
              type: string
              format: uri
              description: Delivery URL (must be HTTPS in production)
            content_type:
              type: string
              enum: [json, form]
              default: json
            secret:
              type: string
              description: HMAC-SHA256 signing secret
        events:
          type: array
          items:
            type: string
          description: Event names to subscribe to
          example: ["push", "landing_request"]
        active:
          type: boolean
          default: true
          description: Whether the webhook is active upon creation
    EditHookOption:
      type: object
      properties:
        config:
          type: object
          properties:
            url:
              type: string
              format: uri
              description: New delivery URL
            content_type:
              type: string
              enum: [json, form]
            secret:
              type: string
              description: New HMAC-SHA256 signing secret
        events:
          type: array
          items:
            type: string
          description: New event names to subscribe to
        active:
          type: boolean
          description: Enable or disable the webhook
...